move modal to its own component file
