+++ src/ui/Card.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const Card = styled.div`
+  background-color: #fff;
+  border: 1px solid #eee;
+  border-radius: 8px;
+  padding: 16px;
+  margin: 12px 16px;
+  width: 100%;
+`;
+
+export default Card;
